Automatically learn analog-control calibration
The default OEHIDEvent class assumes that controller devices will always
use the full range specified by IOKit.  For example, if IOKit maps a
control to the range 0..255, OEHIDEvent assumes that every attached device
will return 0 and 255 when pushed to its physical limits.

Many N64 controller adapters apparently do not behave this way.  They
return less than the full range of values IOKit expects, leading to
characters and vehicles being unable to use their full range of motion.
Link can only tiptoe in Zelda games, for example.

This patch teaches the OEDeviceHandler class to automatically learn the
bounds of each analog axis, individually.  So if the actual physical range
of the X axis of an N64 controller is 49..192, all logical values will be
scaled based on that range instead of 0..255.

This, I think, is tidier than adding a calibration interface.  Why ask the
user to configure what the software can learn for itself?

One limitation is that OEDeviceHandler must relearn its range each time
you start the the emulator.  That means until you've pushed the joystick
through its full range of motion, the scaling will behave in unexpected
ways.  For example, if you push an X axis to the right, the learned range
might be 128..192.  That rightward push will correctly be scaled to +1.0,
but when you return the stick to the center, 128 will now scale to -1.0.
Pushing the stick to the left fixes this by teaching the program that the
minimum is (e.g.) 49 and not 128.  Put simply, the user must push all
analog axes through their full range of motion at least once when the
emulator starts.

There's room for improvement on this "relearn" issue, which I'll address
in a follow-up commit.