Refuse to scale until 8 samples have been learned
This commit fixes the "relearn" problem from 65fcd7af2777.  We now won't
scale a positive value unless a (potentially non-continuous) subsequence
of 8 monotonically increasing values have been observed, and we won't
scale a negative value unless 8 decreasing values have been observed.
While we're waiting for those 8 samples, we simply use the default scaler
that assumes the device returns the full range of values expected by
IOKit.

To put it simply, we will use the default scalar for each direction of
each axis until we have 8 good samples about that direction of that axis.

This learning algorithm can still get confused about the center.  If you
push an axis to the top of its range (100%) but only halfway to the bottom
(25%), then the learned middle will be 62.5% instead of 50%, so 50% will
actually map to a slightly negative value outside the dead zone.  (Link
walks slowly to the left...)  The problem is self correcting as soon as
the user pushes the stick to both ends of the range.

There's still room to improve that.  Instead of learning the middle, we
can assume that each device's actual middle is somewhere near the IOKit
expected middle.  We really only need to learn min and max.