Handle mismatching HID profiles of controllers with same PID & VID when connected to the same instance of OpenEmu.
As a side effect, if this condition happens, the controller bindings won't be able to persist very long (it depends on the last controller being plugged in). At least we don't crash! Apart from the Nintendo stuff which was fixed in 3f00c31, I've never seen two controllers that do this, so I don't think it is important for this case to work flawlessly.