Avoid crashes when decoding OEEvent by not archiving the underlying NSEvent.
NSEvent is considered an user-interface-level API, so it is not safe to use it in a background thread.
Thread-safety related crashes were observed in -[NSEvent eventWithCGEvent:] on macOS Mojave 10.14.6 (18G6032).

The only effective change is that OEEvent no longer forwards invocations to the underlying NSEvent, but no explicit uses of this were found nor observed in the codebase.