chore: Reduce public API surface area
This is preliminary work to explore the possiblity of supporting the
Game Controller framework as an alternative device manager.

**Why?**

We hypothesise: Can OpenEmuSystem be shared on macOS and mobile
platforms?

The desire is that the following APIs are shared:

* bindings,
* bindings management,
* event system (HIDEvent, etc)

The underlying device management would have to change, as IOKit is not
available on mobile platforms. OpenEmuSystem would use the
GameController framework as an alternative and generate appropriate
OEHIDEvent instances. Ideally, clients and game cores would require
none to minimal changes.

To support testing, we should be able to explicitly define a conditional
compiler variable to use the GameController framework vs IOKit on macOS
platforms.

OpenEmuKit would also require changes to disable XPC support when using
a version of OpenEmuSystem that only supports GameController, as there
is no way to pass a device handle between processes to specify which
user is using which controller and there is no unique controller
identifer exposed via the GameController framework[^1]

[^1]: This problem could be solved by adding additional logic to obtain a unique identifier from IOKit on macOS only